requirements.txt
# Stop the server (Ctrl+C)

# Update requirements.txt
cat > requirements.txt << 'EOF'
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==3.0.0
numpy>=1.26.0
pandas>=2.2.0
aiofiles==23.2.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
torch
transformers
scikit-learn
EOF

# Reinstall with compatible versions
pip install --upgrade pip
pip install numpy pandas
pip install -r requirements.txt